Renamed curl_wrapper*.* to rest*.*.
